Adding test in one file
